fixed components and added readmes
