+++ README.md
@@ -1,6 +1,12 @@
 # project-dashboard-rcl
 
-A dashboard that shows various metrics to understand progress of a Bible translation project.
+Set of components that helps to create various metrics to understand progress of projects.
+
+Library have two set of components
+
+- basic UI components : These are indivuadual simple components can be used indivudualy or combined together to create dashbord widgets
+
+- Widgets : Set of widgets created using Basic UI components shows info computed using the data from a JSON
 
 ### Run Styleguidist
