--- src/components/ExampleDashboards/_readme.md
@@ -1,6 +1,0 @@
-# Ag Project Managament DashBoard
-
-DashBorad Examples using the defined components of project-dashbord-rcl
-
-- data will be read from a json file for the called UI Element
-- Example Dashboard components work with a json
