+++ src/components/components/CardLayout/CardLayout.md
@@ -6,7 +6,7 @@ CardLayout component
 import CardLayout from "./CardLayout";
 <div className="flex gap-4 items-end">
   <CardLayout />
-  <CardLayout bgColor="#ff4d4d" width="12rem" height="12rem">
+  <CardLayout bgColor="#ff4d4d" width="14rem" height="14rem" extraClass={"p-3"}>
     <CardLayout />
   </CardLayout>
 </div>;
