+++ src/components/components/_readme.md
@@ -1,1 +1,1 @@
-components of project-dashbord-rcl
+basic UI components of project-dashbord-rcl
