+++ src/components/index.js
@@ -1,2 +1,2 @@
 export * from "./components";
-export * from "./ExampleDashboards";
+export * from "./widgets";
