+++ src/components/widgets/ProgressRateLineChart/ProgressRateLineChart.md
