+++ src/components/widgets/ProgressRateLineChart/index.js
