+++ src/components/widgets/VerificationBarCard/VerificationBarCard.js
@@ -1,12 +1,12 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import PropTypes from "prop-types";
 // import useDateData from "../../hooks/useDateData";
 import CardLayout from "../../components/CardLayout";
 import Label from "../../components/Label";
-import MultiLineChart from "../../components/MultiLineChart";
+import StackedBarChart from "../../components/StackedBarChart";
 import useGetJsonPathdata from "../../hooks/useGetJsonPathdata";
 
-const ProgressRateLineChart = ({
+const VerificationBarCard = ({
   json,
   basePath,
   Head,
@@ -22,27 +22,21 @@ const ProgressRateLineChart = ({
     json,
     currentBook
   );
-  // generate graph data for chart from json
+
   const [graphdata, setGraphdata] = useState([]);
 
+  // generate graph data for chart from json
   useEffect(() => {
     const graphDataTemp = [];
     Object.entries(currentBookdata["months"]).forEach(([key, val]) => {
       const monthobj = { name: key };
-      let sum = 0;
       graphdataKeys.forEach((graphkey) => {
         const elementValueSum = val[graphkey]?.reduce(
           (acc, key) => acc + key,
           0
         );
-        sum += elementValueSum;
         monthobj[graphkey] = elementValueSum;
       });
-      const totalHours = val["TotalHours"]?.reduce(
-        (acc, hour) => acc + hour,
-        0
-      );
-      monthobj["rate"] = sum === 0 ? 0 : sum / totalHours;
       graphDataTemp.push(monthobj);
     });
     setGraphdata(graphDataTemp);
@@ -52,8 +46,8 @@ const ProgressRateLineChart = ({
     <>
       <CardLayout
         bgColor="#ffffff"
-        width="w-auto"
-        height="h-auto"
+        width="auto"
+        height="auto"
         borderRadius="6px"
         extraClass="border-gray-200 border-2 "
       >
@@ -67,16 +61,12 @@ const ProgressRateLineChart = ({
         </div>
         <CardLayout
           bgColor="transparent"
-          width="w-auto"
-          height="h-[15rem]"
+          width="auto"
+          height="15rem"
           extraClass="my-2 mr-2"
         >
-          <MultiLineChart
-            data={{
-              graphData: graphdata,
-              colors: graphColors,
-              dataKeyNames: ["rate"],
-            }}
+          <StackedBarChart
+            data={{ graphData: graphdata, colors: graphColors }}
           />
         </CardLayout>
       </CardLayout>
@@ -84,7 +74,7 @@ const ProgressRateLineChart = ({
   );
 };
 
-ProgressRateLineChart.propTypes = {
+VerificationBarCard.propTypes = {
   /** Json file */
   json: PropTypes.object.isRequired,
   /** Json base path after Projects  - currentBook eg :("projectName/Books") */
@@ -104,6 +94,6 @@ ProgressRateLineChart.propTypes = {
   }),
 };
 
-ProgressRateLineChart.defaultProps = {};
+VerificationBarCard.defaultProps = {};
 
-export default ProgressRateLineChart;
+export default VerificationBarCard;
