+++ src/components/widgets/VerificationBarCard/VerificationBarCard.md
