+++ src/components/widgets/VerificationBarCard/index.js
